+++ inst/train-examples/keras-example.R
@@ -127,7 +127,7 @@ for (i in 1:25) {
 dev.off()
 exp$upload_asset("FashionMNISTResults.png")
 
-exp$log_other(key = "Created by", value = "cometr")
+exp$log_other(key = "Created from", value = "cometr")
 
 exp$print()
 exp$stop()