+++ inst/train-examples/nnet-example.R
@@ -125,7 +125,7 @@ exp$log_html("
 <p>For help on the Comet R SDK, please see: <a href=https://www.comet.ml/docs/r-sdk/getting-started/>www.comet.ml/docs/r-sdk/getting-started/</a></p>
 ")
 
-exp$log_other(key = "Created by", value = "cometr")
+exp$log_other(key = "Created from", value = "cometr")
 
 exp$print()
 exp$stop()